feat: Allow users to disable the suffix range request (#5188)
At present, there are still many storage backends that do not support
the `suffix range` request. I hope to still retain a parameter for users
to control for themselves whether to use `suffix range` for data
reading.

> Why not directly tag the `suffix range` feature for each backend at
the IOClient level?

Because some are in the form of protocols, such as S3, which is a
standard protocol, many different object storage implementations will
attempt to follow the S3 protocol. But, there are also many specific
implementations that do not support the `suffix range` capability. We
cannot directly control at the S3 level whether `suffix range` can be
supported

## Changes Made

1. Allow users to disable the suffix range request because some storage
backends do not support suffix range


## Related Issues


## Checklist

- [x] Documented in API Docs (if applicable)
- [x] Documented in User Guide (if applicable)
- [x] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [x] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)

---------

Co-authored-by: Cory Grinstead <cory.grinstead@gmail.com>
Co-authored-by: Colin Ho <colin.ho99@gmail.com>