perf: Use growable for build side (#5613)
## Changes Made

Currently for joins, all the record batches from the build side are
concated together before the probe, which can lead to significant memory
usage and performance regression.

Instead we should leave them as they are, and we can use the
`RecordBatchGrowable`. The probe side doesn't need to change though.

## Related Issues

This is related to https://github.com/Eventual-Inc/Daft/pull/5160 where
we shifted completely from growable to a `take` approach, which showed a
lot of improvement, but we didn't consider the case where concating the
build side may not be good for memory.

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly