feat: support label_selector specification in ray actor/task creation (#5042)
## Changes Made
When there are multiple worker groups in a Ray cluster, tasks need to be
submitted based on the specific labels of the worker groups.
Therefore, a `label_selector` parameter has been added to the UDF. For
detailed usage, please refer to the Ray documentation at
https://docs.ray.io/en/latest/ray-core/scheduling/labels.html#monitor-nodes-using-labels

<!-- Describe what changes were made and why. Include implementation
details if necessary. -->

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## Checklist

- [ ] Documented in API Docs (if applicable)
- [ ] Documented in User Guide (if applicable)
- [ ] If adding a new documentation page, doc is added to
`docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag @/ccmao1130
for docs review)