fix: fix btree index invalid issue when reading lance for point lookup (#5673)
## Changes Made

**lance:**

https://github.com/lance-format/lance/blob/main/rust/lance/src/dataset/scanner.rs#L1673-L1711

https://github.com/lance-format/lance/blob/main/rust/lance/src/dataset/scanner.rs#L3577-L3595

**daft:**

https://github.com/Eventual-Inc/Daft/blob/main/daft/io/lance/lance_scan.py#L45

`ds.scanner(fragments=fragments, columns=required_columns,
filter=filter, limit=limit)
`

as long as fragments is passed to the lance scanner, the index of lance
will become invalid and the point lookup will be very slow.

Therefore, in `daft`, when it is automatically judged as a point lookup
and indexes have been created for all point lookup columns, avoid
passing fragments to the lance scanner in order to use the lance index
to achieve more efficient query efficiency.


<!-- Describe what changes were made and why. Include implementation
details if necessary. -->

## Related Issues
https://github.com/Eventual-Inc/Daft/issues/5645
<!-- Link to related GitHub issues, e.g., "Closes #123" -->

## How to Test
new UTs.

---------

Co-authored-by: cancai <caican@xiaomi.com>