refactor(arrow-rs): Remove arrow2 Index generics usages  (#5761)
## Changes Made

Arrow-rs doesn't have the notion of an `Index` type like arrow2 does.
Instead, most of the functions either except a UInt64Array or do
internal downcasting from a `dyn Array`. Most of the times, we actually
just force a UInt64Array anyways, like in `argsort`, and most of the
`take` implementations, so I just made it a hard requirement.

Note that I did modify some `take` tests because we no longer support
take on non u64 arrays. If you feel like this would be a regression, I
will add it back by adding a cast in `PySeries::take`