chore: Provide query end state to `RuntimeStatsManager` on query end (#5791)
## Changes Made

Provide an end state (finished, failed, cancelled) to the runtime stats
when the query ends. And only log that there are still active nodes if
it is finished. (it is ok for failed or cancelled to still have active
nodes, as they probably were dropped / cancelled before they called
`finalize_node`)

This is mostly to get rid of any `RuntimeStatsManager finished with
active nodes {2}` logs, as they are not super informative or important
in a failed / cancelled query.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->