refactor: write empty dataframe to parquet/json files via native IO (#5682)
## Changes Made

Currently, if the final dataframe is empty, will use the pyarrow fs to
write the empty parquet/json files, but if the dataframe is not empty,
will use the daft native io to write the data, the write behavior is not
consistent, especially if we want to implement new storage backend, we
need to cover both scenarios.

Secondly, the write empty dataframe to json file is not working
currently, because only parquet and csv file format supported during
write_empty_tabular, the json file format is not implemented via pyarrow
fs.

BTW, there still remains a issue that we can write a empty json, but
cannot read the empty json via `daft.read_json()` since cannot infer the
schema info.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->