feat: Improve model api typing (#5809)
## Changes Made

Types the `**options` param in the model APIs using TypedDicts. Each
model api now has it's own specific TypedDict that indicates the a
subset of the possible options that can be passed, providing better
typing hints for the user and their IDEs.

For most of the model apis, the options include `max_retries`,
`on_error`, and `batch_size`, which are subsequently passed into the
UDF.

Provider specific options can be subclassed, for example the
`OpenAIPromptOptions` is a subclass of `PromptOptions`, and has an
additional field for `use_chat_completions`, which is specific to openai
prompting.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->