feat: Implement retry-after mechanism for model apis and udfs (#5769)
## Changes Made

Adds a `RetryAfterException` that can be raised from a UDF, which will
supply the UDF retry mechanism with a delay to sleep.

This allows us to respect the retry-afters provided in headers from 429
and 503 errors from clients, such as our openai and google clients in
the model apis.

By default, all of our model apis should already be configured with
max_retries = 3, and our exponential backoff has an initial delay of
100ms. It might be worth exploring in the future how we may want to more
easily expose these configurations to users.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->