chore: Fix a minor ambiguity in the README docs (#5830)
## Changes Made

<!-- Describe what changes were made and why. Include implementation
details if necessary. -->

Change "We <3 developers!" to "We ❤️ developers!". I know this is a
rather interesting expression, but some customers have interpreted this
sentence as "Daft has no more than 3 full-time developers", mistakenly
thinking that Daft is a personal project and thus being hesitant to
adopt it in production easily.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

Signed-off-by: plotor <zhenchao.wang@hotmail.com>