fix: Set default ImageMode in decode_image to RGB (#5827)
## Changes Made

Set the default ImageMode for decode_image to RGB. Currently, the
behavior is to infer the image mode at a per-image level, which means we
could have 8-bit and 16-bit images in the same column, which will error
because the underlying datatypes (uint8 vs uint16) are different.

The solution to this is to force decode into a specific image mode. This
PR simply makes that the default.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->