chore: Add basic usage instructions for daft-dashboard in development guide (#5865)
## Changes Made

<!-- Describe what changes were made and why. Include implementation
details if necessary. -->

Add instructions on the basic usage of Daft Dashboard in the development
guide to reduce the learning cost for developers.

Daft Dashboard is currently in the rapid development stage. There will
be a dedicated article or page to introduce its functions and usage in
detail later, but it is more appropriate to include it in the
development guide at this stage.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->

Signed-off-by: plotor <zhenchao.wang@hotmail.com>