chore: bump `mypy` and `ruff` in pre-commit (#5836)
## Changes Made

Bump `mypy` to 1.19.1 and `ruff` to 0.14.10 and resolve pre-commit
errors.

Rationale: my local `ruff` version is much higher than that in the
project `pre-commit`, so I am getting a lot of linter warnings in my
IDE. It is also about time to upgrade lints to utilize newer Python
language features.

## Related Issues

<!-- Link to related GitHub issues, e.g., "Closes #123" -->